+++ test/spec.js
@@ -3495,6 +3495,7 @@ describe('id', function() {
                         '<option value="BR">Brazil</option>',
                         '<option value="CH">Switzerland</option>',
                         '<option value="CL">Chile</option>',
+                        '<option value="CN">China</option>',
                         '<option value="CZ">Czech</option>',
                         '<option value="DK">Denmark</option>',
                         '<option value="EE">Estonia</option>',
@@ -3622,6 +3623,29 @@ describe('id', function() {
             this.$id.val(validSamples[i]);
             this.bv.validate();
             expect(this.bv.isValid()).toBeTruthy();
+        }
+    });
+
+    // #793
+    it('Chinese citizen identification number', function() {
+        this.bv.updateOption('id', 'id', 'country', 'CN');
+
+        // Valid samples
+        var validSamples = ['450202201409072332', '22011219930407001X', '110108601017023'];
+        for (var i in validSamples) {
+            this.bv.resetForm();
+            this.$id.val(validSamples[i]);
+            this.bv.validate();
+            expect(this.bv.isValid()).toBeTruthy();
+        }
+
+        // Invalid samples
+        var invalidSamples = ['999999199304070016', '220112190002290016', '220112199304070019', '999999601017023', '110108999999023'];
+        for (i in invalidSamples) {
+            this.bv.resetForm();
+            this.$id.val(invalidSamples[i]);
+            this.bv.validate();
+            expect(this.bv.isValid()).toEqual(false);
         }
     });